FIX CODE SCANNING ALERT NO. 80: UNCONTROLLED DATA USED IN PATH EXPRESSION (#11)
_Fixes
[https://github.com/private-collaboration-consortium/ckernel/security/code-scanning/80](https://github.com/private-collaboration-consortium/ckernel/security/code-scanning/80)._

_To fix the problem, we need to validate the user input before using it
to construct file paths. Specifically, we should ensure that the `src`
variable does not contain any path traversal sequences or invalid
characters. We can achieve this by checking for the presence of "..",
"/", or "\\" in the `src` variable and rejecting the input if any are
found._
1. _Add a validation function to check for invalid sequences in the user
input._
2. _Use this validation function to validate the `src` variable before
using it in file operations._

Signed-off-by: Pedro S. <pedromiguelfilipesimoes@proton.me>