FIX CODE SCANNING ALERT NO. 245: TIME-OF-CHECK TIME-OF-USE FILESYSTEM RACE CONDITION (#24)
_Fixes
[https://github.com/private-collaboration-consortium/krlean/security/code-scanning/245](https://github.com/private-collaboration-consortium/krlean/security/code-scanning/245)._

_To fix the TOCTOU race condition, we should use file descriptors
instead of file names for file operations. Specifically, we will replace
the `fopen` call with `open` to obtain a file descriptor and then use
`fdopen` to convert the file descriptor to a `FILE*` stream. This
ensures that the file being operated on is the same as the one checked._
*
_[`utils/mkpkg/mkpkg.c`](diffhunk://#diff-04f020a0db76bf6a418de7f2d303deaeee6596fe017eacedb91ab488a45a53ebR8-R9):
Added `#include <fcntl.h>` for file control options._
*
_[`utils/mkpkg/mkpkg.c`](diffhunk://#diff-04f020a0db76bf6a418de7f2d303deaeee6596fe017eacedb91ab488a45a53ebL161-R177):
Replaced `stat` with `open` and `fstat` to check file status and handle
file descriptors more robustly._
*
_[`utils/mkpkg/mkpkg.c`](diffhunk://#diff-04f020a0db76bf6a418de7f2d303deaeee6596fe017eacedb91ab488a45a53ebL234-R245):
Updated the `add_file` function to use `fdopen` instead of `fopen` for
converting file descriptors to file pointers, and added appropriate
`close` calls to ensure file descriptors are closed properly._

Signed-off-by: GitHub <noreply@github.com>