2025.01.27.0820 (#76)
_This pull request includes several changes to the `src/utils/ar/ar.c`
file and a minor update to the `.github/workflows/codeql.yml` file. The
most important changes focus on improving memory management and adding
validation for archive filenames._

*
_[`src/utils/ar/ar.c`](diffhunk://#diff-cbe3270dd16364245b861b64f0b8f885a0c2517823c57f76a91bcbc3c93a8deeL216):
Removed unnecessary null checks before calling `free` for `e->contents`
and `s->name` in the `free_archive` function.
[[1]](diffhunk://#diff-cbe3270dd16364245b861b64f0b8f885a0c2517823c57f76a91bcbc3c93a8deeL216)
[[2]](diffhunk://#diff-cbe3270dd16364245b861b64f0b8f885a0c2517823c57f76a91bcbc3c93a8deeL226)_
*
_[`src/utils/ar/ar.c`](diffhunk://#diff-cbe3270dd16364245b861b64f0b8f885a0c2517823c57f76a91bcbc3c93a8deeR617-R623):
Added validation for the archive filename in the `main` function to
ensure it is safe before proceeding with operations. This validation is
performed both before reading input object files and before writing the
output archive.
[[1]](diffhunk://#diff-cbe3270dd16364245b861b64f0b8f885a0c2517823c57f76a91bcbc3c93a8deeR617-R623)
[[2]](diffhunk://#diff-cbe3270dd16364245b861b64f0b8f885a0c2517823c57f76a91bcbc3c93a8deeR680-R687)_

---------

Signed-off-by: GitHub <noreply@github.com>